requirements.txt
altgraph==0.17.5
cabarchive==0.2.5
cx_Freeze==8.6.3
filelock==3.25.2
freeze-core==0.6.1
lief==0.17.6
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
numpy==2.4.3
packaging==26.0
pefile==2024.8.26
platformdirs==4.9.4
pygame-ce==2.5.7
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
python-msilib==0.5.0
pywin32-ctypes==0.2.3
rich==14.3.3
setuptools==82.0.1
striprtf==0.0.29
textual==8.1.1
typing_extensions==4.15.0
uc-micro-py==2.0.0
wheel==0.46.3
